Refactor Settings
