+++ src/app/components/Settings/index.js
@@ -1,0 +1,3 @@
+import Connection from './Connection';
+
+export default Connection;
